+++ lib/src/model/platform/platform_overwrite_enum.dart
@@ -1,0 +1,6 @@
+enum PlatformOverwrite {
+  android,
+  iOS,
+  web,
+  desktop,
+}