+++ lib/src/widget/touch_feedback/touch_feedback.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:icapps_architecture/icapps_architecture.dart';
+import 'package:icapps_architecture/src/model/platform/platform_overwrite_enum.dart';
 import 'package:icapps_architecture/src/widget/touch_feedback/touch_manager.dart';
 
 const androidDarkTapColor = Color(0x0A000000);
@@ -48,15 +49,14 @@ class TouchFeedBack extends StatelessWidget {
   final Color? shadowColor;
   final ShapeBorder? shapeBorder;
   final Color? androidSplashColor;
-  final bool forceAndroid;
-  final bool forceIOS;
   final bool isAndroidDark;
   final bool isIosDark;
   final bool animateAwait;
   final MouseCursor cursor;
-  final void Function(PointerEnterEvent)? onEnter;
-  final void Function(PointerExitEvent)? onExit;
-  final void Function(PointerHoverEvent)? onHover;
+  final PlatformOverwrite? forcePlatform;
+  final ValueChanged<PointerEnterEvent>? onEnter;
+  final ValueChanged<PointerExitEvent>? onExit;
+  final ValueChanged<PointerHoverEvent>? onHover;
 
   /// Custom touch effect builders will be show on top of the child in a stack
   final List<TouchEffectBuilder> touchEffectBuilders;
@@ -75,8 +75,9 @@ class TouchFeedBack extends StatelessWidget {
     this.shadowColor,
     this.shapeBorder,
     this.androidSplashColor,
-    this.forceAndroid = false,
-    this.forceIOS = false,
+    @Deprecated('Use platform instead') bool forceAndroid = false,
+    @Deprecated('Use platform instead') bool forceIOS = false,
+    PlatformOverwrite forcePlatform = PlatformOverwrite.web,
     this.waitUntilOnTappedFinishesIOS = true,
     this.animateAwait = true,
     this.cursor = MouseCursor.defer,
@@ -84,12 +85,16 @@ class TouchFeedBack extends StatelessWidget {
     this.onExit,
     this.onHover,
     super.key,
-  });
+  }): this.forcePlatform = forceAndroid
+            ? PlatformOverwrite.android
+            : forceIOS
+                ? PlatformOverwrite.iOS
+                : forcePlatform;
 
   @override
   Widget build(BuildContext context) {
-    final isAndroid = (!forceIOS && context.isAndroidTheme) || forceAndroid;
-    final isMobile = isAndroid || (!forceIOS && context.isIOSTheme) || forceIOS;
+    final isAndroid = (forcePlatform != PlatformOverwrite.iOS && context.isAndroidTheme) || forcePlatform == PlatformOverwrite.android;
+    final isMobile = isAndroid || context.isIOSTheme || forcePlatform == PlatformOverwrite.iOS;
 
     Widget touchManager = TouchManager(
       animateAwait: animateAwait,
@@ -104,9 +109,7 @@ class TouchFeedBack extends StatelessWidget {
                 animationController: info.animationController,
                 durationSeconds: info.durationInSeconds,
                 borderRadius: info.borderRadius,
-                rippleColor: isAndroidDark
-                    ? androidDarkRippleColor
-                    : androidLightRippleColor,
+                rippleColor: isAndroidDark ? androidDarkRippleColor : androidLightRippleColor,
               ),
         ],
         ...touchEffectBuilders,